backend/server_3/requirements.txt
flask==2.3.2
flask-cors==4.0.0
torch==2.0.1
numpy==1.24.4
mediapipe==0.10.9
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
